% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_addthresholds.R
\name{rw_addthresholds}
\alias{rw_addthresholds}
\title{Extract Threshold Values}
\usage{
rw_addthresholds(qtr)
}
\arguments{
\item{qtr}{current fiscal quarter, eg "q2"}
}
\description{
Extract Threshold Values
}
