% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_assemble_plots.R
\name{rw_assemble_plots}
\alias{rw_assemble_plots}
\title{Assemble various visuals for country MER performance review}
\usage{
rw_assemble_plots(filepath_msd, opunit)
}
\arguments{
\item{filepath_msd}{filepath to the OUxIM MSD (MSD must be saves as an rds)}

\item{opunit}{operating unit to observe}
}
\description{
Assemble various visuals for country MER performance review
}
\examples{
\dontrun{
rw_assemble_plots("Tanzania", "~/Data/MER_Structured_Dataset_OU_IM_FY17-18_20180921_v2_2.rds") }
}
