% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_gen_nn_target.R
\name{rw_gen_nn_target}
\alias{rw_gen_nn_target}
\title{Append NET_NEW target}
\usage{
rw_gen_nn_target(df)
}
\arguments{
\item{df}{data frame to add NET_NEW Targets onto}
}
\description{
Append NET_NEW target
}
