% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_identify.R
\name{rw_identify}
\alias{rw_identify}
\title{Identify underperforming partners}
\usage{
rw_identify(df, threshold = NULL)
}
\arguments{
\item{df}{dataframe to use, needs to contain cumulative and targets}

\item{threshold}{what level counts as under achievement? eg Q3 = 0.65}
}
\description{
Identify underperforming partners
}
