% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_muliplot.R
\name{rw_multiplot}
\alias{rw_multiplot}
\title{Multiplot function from Cookbook for R}
\source{
Cookbook for R http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/
}
\usage{
rw_multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{objects to be passed in of}

\item{plotlist}{list of graphs to plot}

\item{cols}{how many columns in layout}

\item{layout}{matrix form layout}
}
\description{
Multiplot function from Cookbook for R
}
