% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_output_mech.R
\name{rw_output_mech}
\alias{rw_output_mech}
\title{Prepare Multiplot for each partner's underperforming mechanism}
\usage{
rw_output_mech(df, ind, mech)
}
\arguments{
\item{df}{data frame to use}

\item{ind}{indicator, eg "HTS_TST"}

\item{mech}{implementing mechanism, eg "21212"}
}
\description{
Prepare Multiplot for each partner's underperforming mechanism
}
