% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_pairdown.R
\name{rw_pairdown}
\alias{rw_pairdown}
\title{Indicator selection}
\usage{
rw_pairdown(df, opunit)
}
\arguments{
\item{df}{data frame to review}

\item{opunit}{operatingunit}
}
\description{
Indicator selection
}
