% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_plot_achv_age.R
\name{rw_plot_achv_age}
\alias{rw_plot_achv_age}
\title{Plot age achievement for a given indicator}
\usage{
rw_plot_achv_age(df, ind, mechid = NULL)
}
\arguments{
\item{df}{data frame to use}

\item{ind}{indicator to graph}

\item{mechid}{if interested in a particular mechanism, eg "82182"; default = NULL}
}
\description{
Plot age achievement for a given indicator
}
