% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_plot_achv_agency.R
\name{rw_plot_achv_agency}
\alias{rw_plot_achv_agency}
\title{Graph agency comparison}
\usage{
rw_plot_achv_agency(df)
}
\arguments{
\item{df}{dataframe to use}
}
\description{
Graph agency comparison
}
