% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_plot_achv_hts.R
\name{rw_plot_achv_hts}
\alias{rw_plot_achv_hts}
\title{Plot achievement for HTS modalities}
\usage{
rw_plot_achv_hts(df, ind)
}
\arguments{
\item{df}{data frame to plot}
}
\description{
Plot achievement for HTS modalities
}
