% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_plot_achv_usaid.R
\name{rw_plot_achv_usaid}
\alias{rw_plot_achv_usaid}
\title{Plot achievement for USAID vs National}
\usage{
rw_plot_achv_usaid(df)
}
\arguments{
\item{df}{data frame to plot}
}
\description{
Plot achievement for USAID vs National
}
