% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_plot_mech_table.R
\name{rw_plot_mech_table}
\alias{rw_plot_mech_table}
\title{Plot a table of all operating unit's mechanisms and their indicator achievement}
\usage{
rw_plot_mech_table(df)
}
\arguments{
\item{df}{}
}
\description{
Plot a table of all operating unit's mechanisms and their indicator achievement
}
