% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_plot_trend.R
\name{rw_plot_trend}
\alias{rw_plot_trend}
\title{Plot mechanism's indicator quarterly trend}
\usage{
rw_plot_trend(df, ind, mechid, target = FALSE)
}
\arguments{
\item{df}{data frame to use}

\item{ind}{indicator plot}

\item{mechid}{mechanism of interest}

\item{target}{graph target data? default = FALSE}
}
\description{
Plot mechanism's indicator quarterly trend
}
