% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_prep_achv_age.R
\name{rw_prep_achv_age}
\alias{rw_prep_achv_age}
\title{Setup table for age graphing}
\usage{
rw_prep_achv_age(df, ind, mechid = NULL)
}
\arguments{
\item{df}{dataframe to use}

\item{ind}{indicator whose age bands you want to graph}

\item{mechid}{if you want to graph a particular mechanism, eg "82818", default = NULL}
}
\description{
Setup table for age graphing
}
