% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_prep_achv_agency.R
\name{rw_prep_achv_agency}
\alias{rw_prep_achv_agency}
\title{Prepare table for use with the agency comparison graphic}
\usage{
rw_prep_achv_agency(df)
}
\arguments{
\item{df}{dataframe to use}
}
\description{
Prepare table for use with the agency comparison graphic
}
