% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_prep_mech_table.R
\name{rw_prep_mech_table}
\alias{rw_prep_mech_table}
\title{Prep table for achievement by indicator and mechanism}
\usage{
rw_prep_mech_table(df)
}
\arguments{
\item{df}{dataframe to use}
}
\description{
Prep table for achievement by indicator and mechanism
}
