% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_prep_trend.R
\name{rw_prep_trend}
\alias{rw_prep_trend}
\title{Reshape selected mechanism/indicator data long}
\usage{
rw_prep_trend(df, ind, mechid)
}
\arguments{
\item{df}{data frame to use}

\item{ind}{indicator selection}

\item{mechid}{mechanism to filter to}
}
\description{
Reshape selected mechanism/indicator data long
}
