% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_pull_threshold.R
\name{rw_pull_threshold}
\alias{rw_pull_threshold}
\title{Extract Threshold}
\usage{
rw_pull_threshold(quarter, lvl)
}
\arguments{
\item{quarter}{current fiscal quarter}

\item{lvl}{"low"; "med", "good"}
}
\description{
Extract Threshold
}
