% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_save.R
\name{rw_save}
\alias{rw_save}
\title{Save visuals}
\usage{
rw_save(folderpath, ou, pd, plotname = NULL, mech = NULL, ind = NULL,
  w = 9, h = 3.75)
}
\arguments{
\item{folderpath}{path to folder where you want this output stored}

\item{ou}{operatingunit}

\item{pd}{period}

\item{plotname}{name of the visual}

\item{mech}{mechanism id (character),, default = NULL}

\item{ind}{indicator, default = NULL}

\item{w}{width (in), default = 9}

\item{h}{height (in), default = 3.75}
}
\description{
Save visuals
}
