% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_summarize.R
\name{rw_summarize}
\alias{rw_summarize}
\title{Summarize numeric data}
\usage{
rw_summarize(df, ...)
}
\arguments{
\item{df}{data frame to aggregate}

\item{...}{list of variables to group collapse/aggregate on}
}
\description{
Summarize numeric data
}
