% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_calc_achievement.R
\name{rw_calc_achievement}
\alias{rw_calc_achievement}
\title{Calcuate achievement}
\usage{
rw_calc_achievement(df, curr_fy)
}
\arguments{
\item{df}{dataframe to add achievement onto}

\item{curr_fy}{current fiscal year, eg 2018L}
}
\description{
Calcuate achievement
}
