% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_compile.R
\name{rw_compile}
\alias{rw_compile}
\title{Setup dataset prior}
\usage{
rw_compile(filepath, opunit)
}
\arguments{
\item{filepath}{filepath to the MSD (rds file)}

\item{opunit}{operatingunit}
}
\description{
Setup dataset prior
}
