% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_plot_achv_hts.R
\name{rw_plot_achv_hts}
\alias{rw_plot_achv_hts}
\title{Plot achievement for HTS modalities}
\usage{
rw_plot_achv_hts(df, ind, agency = NULL)
}
\arguments{
\item{df}{data frame to plot}

\item{ind}{indicator to specify, "HTS_TST" or "HTS_TST_POS"}

\item{agency}{can specify agency to filter to, eg "USAID" or "CDC", default == NULL}
}
\description{
Plot achievement for HTS modalities
}
